Refactor: useModal 프롭 타입 개선
1. 두 번째 인자에 null 받을 수 있게 변경
2. 두 번째 인자 ModalHandler 타입 가질 수 있도록 변경
