+++ hooks/use-modal/index.ts
@@ -12,6 +12,6 @@ export { default as ModalListProvider } from './provider/ModalListProvider';
 
 export { default as Modal } from './modal-component/Modal';
 
-export { default as ModalList } from './modal-component/Modals';
+export { default as ModalList } from './modal-component/ModalList';
 
 export default useModal;
